pyproject.toml
[project]
name = "agenticstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "crewai>=1.11.0",
    "crewai-tools>=1.11.0",
    "python-dotenv>=1.1.1",
]
